Move the chunk packing tests into their own file
How a send is packed rather than how it is delivered: the packer and its
event-count rider, the size ceiling, what gets stripped from a record before
it goes on the wire, the compression figures, and the receive-side reassembly
of counts that spanned chunks.
